Move timer gate to common package
